# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Pursuit Bank Mortgage Loan System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get loan product details
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoanProduct"

  # ── Documents ──────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    LoanProduct:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        productType:
          type: string
        description:
          type: string
        minLoanAmount:
          type: number
        maxLoanAmount:
          type: number
        minCreditScore:
          type: integer
        maxLtv:
          type: number
        maxDti:
          type: number
        term:
          type: integer
        isActive:
          type: boolean
        features:
          type: array
          items:
            type: string
        eligibilityNotes:
          type: string
      required:
        - id
        - name
        - productType
        - description
        - minLoanAmount
        - maxLoanAmount
        - minCreditScore
        - maxLtv
        - maxDti
        - term
        - isActive
        - features
        - eligibilityNotes
...